+++ tests/testthat/test_addTAPD.R
@@ -62,12 +62,14 @@ test_that("Repeated doses without samples",{
     doses.1 <- NMcreateDoses(TIME=c(0,3,12,24),CMT=1,AMT=400e3)
     doses.2 <- NMcreateDoses(TIME=0,addl=list(ADDL=4,II=6),CMT=2,AMT=20)
     doses.all <- rbind(doses.1,doses.2,fill=TRUE)
-    doses.all <- egdt(doses.all[,!("ID")],data.table(ID=1:2))
-    dat.all <- addEVID2(doses.all,time.sim=seq(0,26,by=2),CMT=3)
+    doses.all <- egdt(doses.all[,!("ID")],data.table(ID=1:2),quiet=TRUE)
+    ## dat.all <- addEVID2(doses.all,time.sim=seq(0,26,by=2),CMT=3)
+    dat.all <- doses.all
     dat.all <- NMorderColumns(dat.all)
     dat.all[,ROW:=.I]
 
-
+    #setorder(dat.all,ID,TIME)
+    
     res <- addTAPD(dat.all[EVID==1])
     ## addTAPD(
     ##     NMexpandDoses(dat.all[EVID==1])